+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Fixes
+- Made nodepool label update on pods synchronous in the binding process
+
 ## [v0.6.0] - 2025-06-16
 
 ### Changed